refactor: 미사용 훅 제거
- api 파일 이름 변경
