+++ src/utils/hooks/useImagesUpload.ts
@@ -1,4 +1,4 @@
-import { uploadFile } from 'api/uploadFile/Uploadfile';
+import { uploadFile } from 'api/uploadFile';
 import { useRef, useState } from 'react';
 import showToast from 'utils/ts/showToast';
